Add cast to short to silence compiler warning, done in more places already to calls to cond_init
    * ACE/ace/OS_NS_Thread.cpp:

+++ ACE/ace/OS_NS_Thread.cpp
@@ -2553,7 +2553,7 @@ namespace {
       {
         const int result = attributes ?
           ACE_OS::cond_init (&evtdata->condition_, *attributes, name, arg) :
-          ACE_OS::cond_init (&evtdata->condition_, type, name, arg);
+          ACE_OS::cond_init (&evtdata->condition_, (short) type, name, arg);
 
         if (result != 0)
           return result;
